Dependency issue between nh and interface(type:Local Gateway) ksync entries.
Gateway Interface(Local/Gateway) is dependant on physical interface.
This interface used to be accepted by ksync even when parent physical interface
is not present. In absence of same vrouter will reject the config of all
dependant ksync like NH, route and mpls. However when parent gets populated none
of the dependants are re-visited as they have been marked as resolved.

Solution is to ignore gateway interface notification till it is populated with
parent physical interface. This will make sure that all dependants are not
rejected in absence of same and not repopulated.

Fix is in same line as of Vlan sub interface.

Change-Id: I47c7048bd84da8ba68c1700f793487f8842f2c4a
Closes-bug: #1770467