Use tbb::atomic<uint64_t> counters in IFMapChannel class
Some of the counters in IFMAP module are accesed by multiple threads.
Hence use tbb::atomic<> to ensure that the counter access is always
thread-safe.

This is especially critical for sequence_number_ which is used to find
if IFMAP Configuration cached in memory is older/newer than what is
newly received. This counter is incremented in main thread, which does
not fall under TashScheduelr policy control as well.

In theory, if this counter falls under non 64-bit aligned address space,
then concurrent access from multiple threads can lead to bad scenarios.
From the core I see that sequence_number_ offset is at 0x268 bytes in
IFMapChannel object, which is indeed not 64 bit aligned.

e.g. stale timer cleanup timeout callback determines that an entry is
old or new based on sequence number. If this number is modified from
the main thread at the exact time when read, and if the read value
happens to be corrupted, then potentially the stale timer callback can
end-up deleting valid configuration.

From the core, it seems BGPRouter object being accessed is already
deleted. Hence the process crashes in boost graph library vertex/edge
access code.

Since the scenario mentioned above wrt concurrent counter access is
less likely to happen, the bug shall be marked as partially fixed.

Separete UTs to conver stale-timeout code are is planned for a separte
commit in the upcoming weeks.

Change-Id: I0422da856a643b3a556f2411a87462d38d9545a6
Partial-Bug: 1770286