PBB gateway interface classified with incorrect device type.
Device type remote vmi with vlan maintains a dependency on parent physical
interface so that its download to vrouter is not done till parent is seen.
Since pbb gateway was wrongly classified as local gateway it used to miss the
dependency check.

Fix is to clasify pbb gateway interface as remote vm with vlan.

Change-Id: I47c7048bd84da8ba68c1700f793487f8842f2c4a
Closes-bug: #1770467