Exception incorrectly handled when 'type' is missing
Description: We have an object with no type but fqname.
https://bugs.launchpad.net/juniperopenstack/+bug/1630130 and the
 fix indicate that objects of that definition may exist. As part
 of the fix, those objects are skipped while reading them. But
when they are part of refs of another object and that object
while rendered, would read the ref object fqname. While reading
the fqname, we fetch the type and cache the type. This caching
of ‘type’ of object caused this exception. This exception was
handled outside the object read, and that resulted in skipping
all the VMI list, as VMI is the object that is rendered having
a ref of VM which is an offending object. This resulted in not
publishing building VMI related graph and not publishing. As
mentioned, it is not clear why those objects exist in first hand,
 but they may exist. Though I don't like, this fix aligns with
the fix done for 1630130.

Change-Id: I07f927477b6abce7849f231bdd47bb726c255bae
Closes-Bug: 1771879