Coverity issues in VmFlowRef constructor
Coverity reports the following issues in the VmFlowRef(const VmFlowRef &rhs) constructor:

Using uninitialized value this->fd_(59780)
Non-static class member fd_ is not initialized in this constructor nor in any functions that it calls.
Non-static class member port_ is not initialized in this constructor nor in any functions that it calls.
Non-static class member flow_ is not initialized in this constructor nor in any functions that it calls. (59794).

Change-Id: Iaf29a87a0adc4746559a1ca9ddf3ee4fc6e41136
Closes-Bug: #1773354