schema_transformer: don't mark LR RTs as stale
When an RT from a LR is linked to RI it is marked as
stale when RIs are initialized.

At the end of schema init, process_stale_objects removes
all stale RTs and LR_RT/RI link is lost.

This patch check if some LR is connected to the RI and
remove the LR RT from the RT stale list of the RI.

Change-Id: I783b3f24421596f0a55478765a0081ae73350e44
Partial-Bug: #1588204