schema_transformer: evaluate LRs at init
When an RT from a LR is linked to RI it is marked as
stale when RIs are initialized.

At the end of schema init, process_stale_objects removes
all stale RTs and LR_RT/RI link is lost.

This patch add an evaluate method to LogicalRouterST so
that LR RTs are removed from RIs stale_route_targets list.

Change-Id: I53ae4430689a0a88ca4fe48c78d474d883f21d3d
Partial-Bug: #1588204