Modify LogicalNetworkVirtualNetwork type as a property of LogicalRouter
Earlier, it was defined as a property of the link between LR and VN.
But this complicates logic during LR/VR creation in api-server.

Modify the schema to denote the link attribute as a property of the
logical-router object.

Modify agent code to retrieve the information from LogicalRouter
object instead from the link attribute of LogicalRouterVirtualNetwork.

Change-Id: Id9d6537f78a32691a7d584bf2cd1e981801a2022
Depends-On: I0164ec246f35c57667357784968ee7c80ff1f63d
Partial-Bug: 1774546