FMG tree not present in agent.
During notification walks there are two types of walks - notify all routes
and notify multicast. The first walk also creates state which second one does not.
In absence of state agent will not add fmg tree as it will see no state and assume
it never subscribed for it. Two ways to fix this issue- Add state creation in multicast route walk
or use notify all routes walk for multicast as well. First option is bit complex
as state addition is not limited to multicast routes but every vrf as well. Hence it will be no
more a multicast walk. So its better to replace it with peer notify walk as its tested and
handles state creation on each route and vrf.

Closes-bug: #1766185

Conflicts:
	src/vnsw/agent/controller/controller_peer.cc
Change-Id: I9badbe203754c07ffadc9163faaaef109d8451f9