Add multivendor capabilities to device discovery
This patch implements multivendor capabilities for the device discovery
process of fabric management. Once the approach is ironed out and this
patch is merged, a future patch will build on this to complete full
multivendor fabric management capabilities.

Summary of changes:

- If existing SNMP classification method fails, the `netmiko` library's
  "autodetect" function is used to determine vendor. This also allows us
  to determine the appropriate NAPALM driver to use for future
  multivendor fabric management functionality.
- Added mapping for the netmiko and NAPALM drivers into group_vars
- Changed the variable `oid_mapped` to `device_info` to be more clear
  about it's purpose.

Please refer to MULTIVENDOR.md in this patch for detailed information
about the approach, and how to test

Change-Id: Ie585befdb1f2756cae86b2e6c0fc67ecacaa3f3c
Signed-off-by: Matt Oswalt <matt@keepingitclassless.net>