Persist operational state for port types of type REMOTE_PORT
Fix issue of port going to oper-down state on vrouter-agent
restart.

vrouter-agent already stores port info in files which are used
to restore ports on vrouter-agent restart. Port-state was not
being stored, because of which default state of oper-down was
used on vrouter-agent restart. This is required only for port
types of REMOTE_PORT because they don't have corresponding tap
interface is OS to pick oper state.

Change-Id: I54144b82af17950ad8bb0c105d9e2ad10e715701
Closes-Bug: #1776876