[config] In draft mode don't remove ref on deleted resources
For a reason I don't remember in draft mode, when API server clones a
resource and puts is in pending delete state, it removes all its reference
to any other resources.
That patch keeps that references in the clone draft version and lets the
delete call run during a commit or a discard action takes care of them.

Change-Id: I1fc286768b359d495c3027ba1bab9bac5a1983c7
Closes-Bug: #1774038