Do not access connection from cleared session
We enqueue XMPP messages to a queue and process. However, this enqueue
is done off IO Reader task. But from main thread, tcp event can cause
OnSessionEvent() which could potentially enqueue TcpClose() to the state
machine ahead of the event that processes the message itself

This bug possibly got exposed due to the change in timings due to commit
8fef2aa03522465f311ff3bdf11bfcdc7078366e for bug fix 1643486

Hence drop processing xmpp message if session is not associated with the
connection. We alaready delete the session through an event in order to
serialize access to session pointer. Tcp also protects TcpSession using
intrusive pointer TcpSessionPtr.

Change-Id: Ib3b49f6102440e582def20bff134690f6e29462a
Closes-Bug: 1649195