Fix a potential race condition in WhereQuery::subquery_processed
When multiple subqueries are created for a where query, all the
subqueries run in parallel and calls subquery_processed when done.
In subquery_processed, the results of the subqueries are copied into
a vector that is a member of the WhereQuery. When the results from
all the subqueries are recieved, necessary set operations are done and
the WhereQuery processing is done. Althoght copying from subquery to the
shared vector part was inside a lock, the check whether this is the last
subquery or not wasn't. This means multiple subqueries can think they were
the last. This race condition could lead to many unwanted scenarios. A
couple of them are -

1) first subquery thinks it is the last, goes ahead with the subsequent processing,
   copies the result into another vector to be passed to the next stage, and frees
   the current vector. The second subquery thinking it was the last tries to access
   the freed vector  and cores.

2) two subqueries thinking they were the last, goes ahead with all the processing in
   parallel. When all the stages are done, the first one frees all the datastructures.
   Second one also tries to free the same datastructures when it is done and cores.

Change-Id: I93eda6ac581f31a9cd91b1d9cc4fd6061bf388e3
Closes-bug: #1769865
Closes-bug: #1759719