[config] Remove RDBMS dead code
An RDBMS backend was introduced 2 years ago [1] but is not maintained
anymore in flavor to a new project [3] which aims to re-write all the
config side of Contrail in Go based on an RDBM system.
That patch removes that dead code and also lightens the unit test suite
(reduce from 599 to 363). It keeps the work done for the RDBMS support
to provide a generic DB interface (there is still no clear interface
defined like it is recommended in that spec [3]).

[1] https://bugs.launchpad.net/juniperopenstack/+bug/1628726
[2] https://github.com/Juniper/contrail
[3] https://github.com/Juniper/contrail-specs/blob/master/pluggable_services.md

Change-Id: If2d71aadec2777c716fe9e9ae599149d251b47c3
Closes-bug: #1779141