Don't send ARP probes
When we leave the SPA (source IP) field empty in an ARP request, the
request begins to be an ARP probe. When container tries to assign
itself an IP address and receives an ARP probe, it thinks that somebody
else has that address already assigned and the address is marked as a
duplicate. To stop this behavior we should send ARP requests from normal
IP address. The idea to use service address was taken from
ArpEntry::SendArpRequest.

Change-Id: Iceec678ed125664711e1c13714514290b37e7c96
Partial-Bug: #1737177