Handle configuration updates with null value as deletes
In IFMAP configuration scheme, there is no such thing as null. However,
in the cassandra db json based notation, value can be "null" which
means that the associated value does not exists (any more)

Current code used to entirely drop such properties with value as "null".
This poses problem as the ifmap db goes out of sync with the config db.

Modify the code to treat such null values as deletes instead.

1. Tested in Heat setup and verified that service-chains get recreated
   correctly
2. Verified that some of the large scale customer DBs can be loaded
   correctly
3. Pending -- Specific unit test and also to evaluate the impact of this
   on property refs.

Change-Id: I45a8411e686bae9e33c5f33e995f19e075c6b835
Partial-Bug: 1782028