Vrouter name lookup with non-FQNAME.
Due to limitation in our provisioning scripts, vrouter may not always
register with controller using FQDN on the host where its running. On
the other hand, when a VM is scheduled by nova, the hostname of the VM
can very well be populated with FQDN of the host.

This causes a problem in Kube-manager which tries to vrouter-object by
comparing hostnames from VM object with the name of the vrouter object.

This issue should ideally be addressed in provisioning.But due to other
complexities, that fix may have to wait.

Till then, this code change will look for vrouter with both FQDN and
non-FQDN.

Change-Id: I7ba5eef2139b201fa890cf2332d3a670412db28d
Closes-Bug: #1784727