Handle exception at bgp_router object deletion before RI
Closes-Bug: 1778501
The bug shows off when the customer was trying to create virtual network, VMI and BGPaaS objects and remove them using script. There is an exception occasionally happened at deletion stage, during  the deletion of ‘bgp_router’ object, a children of ‘routing_instance’ object. We found there were two deletion requests for this ‘bgp_router’ object, one came from schema and another came from ‘internal_request_delete’ which trying to remove the children object before removing ‘routing_instance’ object. These two deletion requests come out of order and sometimes ‘HttpError’ or ‘NoIdError’ exception happen, which cause the deletion request for ‘routing_instance’ object not being issued.
To solve this problem, we try to catch the exception when running ‘internal_request_delete’ of ‘bgp_router’ object. If there is an exception happened, we raise the exception when the object is not being removed.

Change-Id: I78a356f622188826e3fb469310a352322467acc5