vrouter-agent: Qos config is not persistent across reboots
Simplifying the previous fix.
Instead of resyncing explicitly for the global Qos, properly handle
the vmi of type Vhost in vm_interface_request.cc by checking if the
config needed to be applied for global_qos. global_qos is applied for
vmi of type VHOST and not for other types.

Change-Id: Ib1df71d9b480b78919cf390868947b88af55fa42
Closes-Bug: 1775831