Bgpaas port range is not updated in running time
when we config the Bgpaas port range, it doesn’t take effect unless we restart
the schema container. The range doesn’t be updated during the configuration, because
'self._alloc_list' is not updated in 'reallocate' function.

Also add the import of 'ResourceExhaustionError', which shows off in
'create_bgp_router' function.

After adding the update to alloc list, we found out an error in ip allocation 
test case. The current test case has a shrunk error, which was not triggered
before because the alloc list was not updated, so it always compare the new
alloc list to the original one created at init. The new change is testing
adding a range, check the delete range and overlap error.

Change-Id: Ib5c3caa99f100c04b41b81382831eec7e305ce9e
Closes-Bug: #1778501