Bgpaas port range is not updated in running time
when we config the Bgpaas port range, it doesn’t take effect unless we restart
the schema container. The range doesn’t be updated during the configuration, because
'self._alloc_list' is not updated in 'reallocate' function.

Also add the import of 'ResourceExhaustionError', which shows off in
'create_bgp_router' function.

Change-Id: Ib5c3caa99f100c04b41b81382831eec7e305ce9e
Closes-Bug: #1778501