Fixes the core file collection issue in nodemgr
The current implementation gets list of core files in
/var/crashes/ periodically and deletes the core files
for the process it is monitoring. nodemgr decides if a
core file is owned by it or not based on the process name
in the core file. However there is a limitation on the length
of the process name string in the core file(14 characters).
For many of the applications that nodemgr monitors, this is
not valid (eg., contrail-vrouter-agent, contrail-device-manager)
This can result in situations where nodemgr could delete
incorrect core files; so this has to be PID based as opposed
to process name based.
The core files to process association happens when ever a process
goes down.The files are actually deleted in the timer event.
Closes-Bug: 1790734

Change-Id: Id3bbade0906f593d8ccd5973959c5129be4ae56b