Fix vDNS responding on Windows
Sometimes agent didn't respond to VMs' DNS queries
in vDNS mode. It was related to memory with the response being freed,
before Pkt0Interface::WriteHandler have been called.
With this fix, we will ensure that the memory won't be
freed until the handler of async_write function is called.

Change-Id: I07734c22be4134fc22ad1625064d7afb6a3391b7
Partial-Bug: #1737177