Implement default DNS mode on Windows
In default DNS mode VMs get list of DNS resolvers
from OS on which vRouter agent is running.
Right now, this list is read from /etc/resolv.conf file,
so it doesn't work on Windows.

To get the list of known DNS servers on Windows, we have to use
"iphlpapi" module, which is IP helper API.

This commit will logically divide implementation of building
the list of DNS resolvers in matter of OS the agent is running on.

Change-Id: Iad1b1dcd2237a4f005c8d3207b0ad3e87499326a
Partial-Bug: #1737177