Do not insert already sleeping tbb task back into sleeping list
During testing, it was found that tbb sleeping threads singly linked
list was corrupted and had become circular. This seemingly caused
my_slack count to get permanently stuck at -1, as the sleeping list
traversal would potentially never end.

During testing, using a specific assert, it was confirmed that duplicate
insertion did happen.

Fixed it by modifying the sleeing threads singly linked list into a
doubly linked list and then making sure that a thread if already in
the list is never prepended back as the head of the list.

https://github.com/01org/tbb/issues/86

Change-Id: I3375b8be324245c329a9bd3a1f001a38576f617d
Closes-Bug: #1684993
Depnds-On: I6773eb8dddd849cebb695a59864a9da2ce2faa17