[config] Improve performance on the permission check
If a non admin user lists a resource type, the permission code needs to
fetch permissions of all children, references and back-references of all
listed resource to check if what can be read by the user.
Actual code does one DB read per listed resource to fetch that
permissions. This patch reduces it to on DB read in improve response
times (around 39% improvement on a list port from Neutron API on 3.2.10
deployment with 360 VNs, 2035 VMIs and 742 VMs).

Change-Id: I85500e15a06395b23dae51e85b08bd6dce9cec93
Partial-Bug: #1762820