[config] Improve performance on the permission check
1/ If a non admin user lists a resource type, the permission code needs to
fetch permissions of all children, references and back-references of all
listed resource to check if what can be read by the user.
Actual code does one DB read per listed resource to fetch that
permissions. This patch reduces it to on DB read in improve response
times (around 39% improvement on a list port from Neutron API on 3.2.10
deployment with 360 VNs, 2035 VMIs and 742 VMs).

2/ If the request header 'HTTP_X_DOMAIN_ID' is not set, RBAC permission
check code needs to resolve domain UUID (mostly the default domain) each
time it needs to check a permission. That patch resolve the domain ID if
not set when a request received and set it the headers.

3/ The patch also forces the DB driver to return a deep dict copy of objects
even it comes from the cache to not have to make a copy in the API
server code when it renders it (href and filtering allowed
child/ref/back_ref (obj_view)).

Change-Id: I85500e15a06395b23dae51e85b08bd6dce9cec93
Partial-Bug: #1762820
(cherry picked from commit 7af8daa11ae4f86f1af87ce2d433d51b8cbd0756)