vhost0 interface status is checked before connecting to controllers
Agent can connect to controllers via any interface and start
advertising the routes. As long as vhost0 is up, there is no
issue in connecting and advertising the routes via other interfaces.
If vhost0 is not up, advertising routes will affect the traffic.
Fix is added to make sure that vhost0 is up before connecting
to controllers and vhost0 ip(interface) is used to connect.

Change-Id: I5f0283e2947f34a4b642febaaf0bc91a31252723
Partial-jira-bug: JCB-219276
Closes-bug: #1793024