[config] Fix TypeError issue when commit security
Fix code to not try to split a None object when commit a global draft
address group with a reference to a project scoped firewall rule in
its endpoint 2.
Also permits to use the saùe address group in the both firewall rule's
endpoints before the draft address group is committed.

Change-Id: Iace2995f6b701e87a2fcbdb7a4656775cc639513
Closes-Bug: #1794954