Fix for active TCP flows not transitioning
Active TCP flows not transitioning from Forward to deny when SG
is updated to Deny. The fix changes the existingbehavior on how
TcpAck flows are treated. Instead of allowing the TcpAck flow when
either of forward or reverse policy allows it, drop the flow when
either evaluates to drop. The fix allows the flow state to transition
to drop as is expected, when SG is updated.
Closes-Bug: #1786924

Change-Id: Iec49dc5fbd8e51f4069fea577db32d8995d0fb11
(cherry picked from commit e00b66c05a18ccdca1b953ae230e2b073d5d021e)