vrouter-agent: R4.1 Some of the BFD sessions are not established
When Health check service for BFD is configured, Agent
should establish the sessions between the Source IP and
vmi IP (health_check_ip). Here the source IP is always the
Gateway IP (.1 suffix)

During BFD scaling tests it was noticed that for some of the
BFD sessions, source IP is set to .2 suffix (aka service IP).

From the logs it had been observed that for the problematic
session, the Healthcheckobject update (resync) had not being
called.
Following fix will trigger Resync from VMI if the corresponding
interface source IP is not set.

root cause:
While processing HC instance we look for source ip via vn->ipam
if not present  source ip is unspecified hence the session is
brought up.
Ipam is updated (or added) on vn, vmi holds pointer to vn and doesn’t
look for this change hence HC is not triggered. Introspect however is
attempting to get the source IP  checks for the service_health_check_ip
in the healthcheck instance, in this case is unspecified
(good and bad case), then proceeds to find service ip instead (dns),
which is what is displayed. DNS IP is displayed although actually
src ip is 0.0.0.0

Change-Id: Iae1a610b7c3085efaed999e717e43a11c46ccbb0
Closes-Bug: 1788682