vrouter-agent: Some of the BFD sessions are not established
When Health check service for BFD is configured, Agent
should establish the sessions between the Source IP and
vmi IP (health_check_ip). Here the source IP is always the
Gateway IP (.1 suffix)

During BFD scaling tests it was noticed that for some of the
BFD sessions, source IP is set to .2 suffix (aka service IP).

From the logs it had been observed that for the problematic
session, the Healthcheckobject update (resync) had not being
called.
Following fix will trigger Resync from VMI if the corresponding
interface source IP is not set.

Change-Id: Iae1a610b7c3085efaed999e717e43a11c46ccbb0
Closes-Bug: 1788682