Enforce K8s APS on Virtual Networks.
This change is to enforce Security Policy APS created for K8s at
the virtual-network level, rather than at Project level. This is
so that unrelated virtual-networks created in the same project will
not be affected by this APS.

In the case of Logical Routers, contrail infra creates a virtual-network
in the project where its enabled. Since the APS was being enforced at
project level, it was applied to this virtual network as well, resulting
in functionality breakage. This change addresses this issue by applying
APS on k8s created virtual-networks only.

Change-Id: I19e8876967aa4b40e5aad75b131ee2e206ebeccf
Closes-Bug: #1794412