Make it possible to modify network flags in Horizon
It is not possible to enable and successfully save Shared or
External network flags for existing network on /admin/networking screen.
To resolve this problem we need to correctly proccess
this types of requests. To check that logic works fine
new test was added and one existing was extended.

Related-Prod: TFB-1517
Change-Id: I2a8382e7e0b67b5f9ae95fe526a947296ff2d3e0