[config] fix resource type listing in common vnc_db code
The python common lib used by config services (schema-transformer,
svc-monitor...) uses the 'vnc_cassandra.object_read' to list all resources
for a type. That method requires a list of resource UUIDs to be read but
in case that list contains only one UUID and that UUID does not exists
anymore, the 'object_read' method raises a NoIdError exception instead
returning an empty list.

Change-Id: I743767f1e441fb26f726bc0acdc8aa29e2ef74f9
Closes-Bug: #1798765