Re-enable assert() while reserving labels
There is a suspicious commit as part of
commit 4877e2aff98151d8e5e2e784bae31f3cedf755cd
where the following change was made..

size_t InsertAtIndex(uint32_t index, EntryType entry) {
        size_t size = bitmap_.size();
        if (size == 0 || size <= index) {
            bitmap_.resize(index + kGrowSize, 1);
             entries_.resize(index + kGrowSize);
         }

+ // TODO(prabhjot) need to enable assertion below
+ // currently disabled due to some issue with MPLS
+ // label allocation
+ // index should not be already in use
+ // assert(bitmap_[index] == 1);

This should be called for MPLS label allocation only to reserve
multicast labels for ERMVPN or to block reserved MPLS
labels 0 though 15.
Need to find the root cause if this issue occurs as it could be
anything from memory corruption to calling ReserveLabel twice.
Note that the infrastructure assumes that sanity is maintained
by the caller.
Closes-Bug: 1800532

Change-Id: If630aaf954fb3ce91ad43352d7fb2e367c093dd3