Add Route-Reflector based bgp peerings configuration
In schema-transformer, tweak full-mesh auto-bgp peerings creation to also
cover a scneario in which one or more BgpRouters of type "router" can be
a route-reflector server using non-zero cluster-id.

The logic is basically to create bgp-peering links only betweem control-nodes
and between route-reflector server and clients.

Also ensure that both ends of the peering link are notified when refs are
added between route reflector server and client BgpRouter objects

UTs for this specific piece shall be added shortly and this review will be
amended.

Change-Id: I15cd0160288c66b4e7b33e4e2f4c1fe1c66573f8
Partial-Bug: #1771741