Detect duplicate subscription from agent even when undergoing graceful restart
Normally, if control-node receives duplicate subscription request from agent
for any virtual-network, then the session is closed and a warning message is
also logged. This is not expected behavior from the agent and control-node
data structures are not designed to handle this. This also likely indicates that
some thing else has already done wrong in the data structures in the agent as
well.

This check was relaxed when the agent was under GR. This is because, when the
agent session is under GR, it is expected to receive a duplicate subscription
when the session comes back up. However, we need to ensure that even under GR,
subscription is received only once, after a session comes back up and not
repeatedly instead, as was seen by analyzing the core reported in this bug.

Hence restored the check to ensure that duplicate subscription are detected
when agent session in GR state or otherwise so that control-node does not get
confused within its state as driven by its internal data structures, especially
the bgp membership manager.

Note that in the cores, it was noticed that duplicate subscription was received
for instance default-domain:default-project:ip-fabric:__default__ and
default-domain:default-project:ip-fabric:ip-fabric

Updated GR tests to
1. Subscribe to default-domain:default-project:ip-fabric:__default__ as well
   and advertise routes
2. Send a duplicate subscription to __default__ instance when undergoing GR and
   ensure that the session does flip. Without this fix, test did crash with the
   exact back-trace that was observered in the cores reported in this bug.
   (graceful_restart_flap_some_test2)

Bug has been marked as partial as root cause still needs to be fixed in agent.

Change-Id: Ib325fd412273dc5c823c56ab7cff70af92a4d836
Partial-Bug: 1800739