Core file deletion not happening as expected
In nodemgr, we allow a maximum of 4 core files for the processes
that it manages. When a process crashes, nodemgr gets a  notification
;it then checks the number of core files and if there are more than
4 files, we delete the oldest entry. We maintain a list internally,
core_file_list which stores the core files corresponding to the process
 stored from oldest to the newest;So upon reaching the maximum core limit,
we remove the oldest file and then append the new core file to list.

It is also possible that the core files could be manually deleted.
In such cases, the core_file_list should reflect the changes. The
nodemgr checks the /var/crashes folder every 1 min and updates,
the core_file_list based on what it currently lists.

In current implementation there is a single place of deletion of
core file list. It happens in the timer expiry context alone.
This handles both online and offline deletion of core files.
If the corefile belongs to a daemon that this nodemgr controls
then it will be deleted.

Change-Id: Ibeecd32ad966bcecb9e55dd0f01537c8cba50c0c
Closes-Bug: 1776170