nodemgr core collection issue
The core file name do not contain the whole process name
and its difficult to identify the process name just with
the core file name. Identifying core file based on the
process id will not work help in the case of nodemgr
restart. To address both issues a timer based core
deletion is introduced. After timer expiry, we check list
of core files and if any of it belongs to the daemon managed
by nodemgr, then it deletes the core file (if the core files
is greater than configureed limit) and reports to collector.
Verified the deletion of file in testbed under the following
scenario:
1) HAve nodemgr running and created more than 4 core files
2) Stopped nodemgr, created more than 4 core files and started
nodemgr again

Change-Id: I35b6df12405a41d4aad266a4ac94f8a66dcd543e
Closes-Bug: 1776170