Add Route-Reflector based bgp peerings configuration
In schema-transformer, tweak full-mesh auto-bgp peerings creation to also
cover a scneario in which one or more BgpRouters of type "router" can be
a route-reflector server using non-zero cluster-id.

The logic is basically to create bgp-peering links only betweem control-nodes
and between route-reflector server and clients.

Also ensure that both ends of the peering link are notified when refs are
added between route reflector server and client BgpRouter objects

Change-Id: Ia4c5b5c3e89bab4990c8a4c1973ba74834e3b317
Closes-Bug: #1771741