[config] Allow to specify action in security group policy rule
The Contrail data model permits to define the action to be done on
policy rules set in a security group but the schema-transformer did not
reflect it to access control list it creates. It always set the action
to 'pass'. This patch fixes the schema-transformer to correctly reflect
the policy rule action to access control list and set it to 'pass' if
not set in the policy rule.

Change-Id: I55d43991f45e3392da68038aa10c84d822c99fab
Closes-Bug: #1805662