Handle port add for VMI previously associated with another VM
When a compute node goes down and comes back up, it reads the saved
port information from the port configuration directory and re-adds
the ports.
In a VM motion use case, a new VM is used to bring back the workload
and reuses the VMI the old VM was assiciated with.
We were not handling this new association and hence no VM subscribe
was sent to the CN for the new VM.

To fix this we treat such as case as an internal port delete for the old
association, and an add for the new assiciation. The port configuration
files are updated. This would result in sending the CN an unsubscribe
for the old VM and a subscribe for the new VM.

Change-Id: I12dee38b8e2abbbf635ced6985543521908fc214
Closes-bug: 1786240