Optimize ifmap provision on contrail-api startup
This patch fixes long contrail-api startup time on big databases:

On 10000 VMs,
without patch: 18 minutes 14 seconds (+-20 seconds)
with patch: 2 minutes 40 seconds (+-10 seconds)

This patch:

* changes loglevel of ifmap provision time to SYS_INFO
* removes unnecessary conversion from str to unicode and back
* removes dead code

About dead code:

In line 716 of vnc_ifmap_client.py there is an error:

    old_metalist.append(self_metas[meta_name])

instead of:

    if id2 in self_metas[meta_name]:
        old_metalist.append(self_metas[meta_name][id2])

So old_metalist contains dicts instead of Metadata.

As a consequence, if request != old_request later in the code is always true.

Instead of fixing the bug, we decided to remove unneeded check
and redundant work associated with it.

Closes-bug: 1807379
Change-Id: I1826b72363f57b2ffdfeff305a12f7a58e41e293