Retry to create resource with only fq_name changed
When resource is created in vnc_openstack, and there is already
resource with given fq_name, it should retry with only fq_name changed
(uuid unchanged). If it fails, only then it should retry with new uuid.

This change is necessary to properly match resources between openstack
and Tungsten Fabric using networking-opencontrail plugin.

closes-jira-bug: NTF-53
Change-Id: I223456a1deb0ea0a6cc066d5406d6723b8676536