[Config] Add support to log remote IP address
As per the current implementation, WEBUI host IP is sent as
remote IP in the logs sent to Analytics.
This change will add actual remote IP address of the end user
when sending the log information to Analytics.
To do this, a WEBUI change will be done seperately. This code
review is to handle the changes in vnc_cfg only.

Change-Id: Ibfd15e298740956695949453996e37863a667af5
Closes-Bug: #1808574
Closes-Jira-Bug: JCB-189711