Move default targets to ci_unittests.json
This change will allow us to run the unit tests
target by target using the new testrunner. Scons
does not provide a way to easily find all the aliases
which are aliases of another alias. This will allow us
to retrieve a list of targets to run with contrail-unittests-gather.rb

Change-Id: I8db46e709c059cde22cd2f2c28b0a09731918836