Active TCP flows not transitioning upon SG attach
When updating reverse flow as part of reflexive action, consider
the rflows calculated as part of the forward flow. Earlier code
copied the action from fwd flow; now if the rflow entries are
available, use them. This avoids setting a rflow to Forward, which
was earlier copied from fwd_flow, when rev-SG and rev-outSG dictate
Deny.
Closes-Bug: #1786924

Change-Id: Ie1dfac6d5c113f4243e7f55ce0da9582deef805e