Active TCP flows not moving to deny upon SG attach
The bug manifested coz we reflexively update a reverse flow,
when a forward flow is updated. We call UpdateReflexiveFlow()
for reverse flow whenever a forward flow is updated.
After the SG attach(allows only egress) to Server port, the
forward flow moves to 'D' and reverse flow to 'H'. When TcpAck
from server port hit the reverse flow, flow is reevaluated and
it becomes F(since egress is allowed and this is TcpAck). And
this reverse flow becomes forward and previous forward flow(which
is 'D') becomes reverse. Now the new reverse flow is updated
reflexively and copies the sg_action_summary from Forward flow to
be toggled to 'F', which leads to the reported problem. This will
allow back traffic from client to server.
Fix here has been to not reflexively update the reverse flow for
forward flows, triggered because of TcpAck packets. This allows the
keepalives from server to work, while allowing the C->S flow to be
dropped.
The change has been made in ResyncFlow since this is called first
for reverse flow and then for the forward flow(expecting the reverse
flow to be updated again as part of the ResyncFlow code).

Closes-Bug: #1786924

Change-Id: Ie1dfac6d5c113f4243e7f55ce0da9582deef805e